Add primitive CPU mapper
This commit adds a primitive CPU mapper and updates the tests.
In the current form, the context is not being used to gist and
conditionals on parameters of the form `if (M == 40) { ... } `
are needed.
So this PR also adds the `emitIf` function to generate the proper
LLVM IR and let the tests pass.